feat: Add base class
