+++ gym_simplifiedtetris/envs/simplified_tetris_binary_env.py
@@ -4,10 +4,10 @@ import numpy as np
 from gym import spaces
 
 from ..register import register
-from .simplified_tetris_base_env import SimplifiedTetrisBaseEnv
+from .simplified_tetris_standard_env import SimplifiedTetrisStandardEnv
 
 
-class SimplifiedTetrisBinaryEnv(SimplifiedTetrisBaseEnv):
+class SimplifiedTetrisBinaryEnv(SimplifiedTetrisStandardEnv):
     """
     A class representing a custom Gym env for Tetris, where the observation space
     is the binary representation of the grid plus the current piece's id.
