+++ requirements.txt
@@ -1,10 +1,9 @@
+numpy==1.19.5
 tqdm==4.51.0
-matplotlib==3.4.2
+gym==0.18.0
 imageio==2.6.0
 stable_baselines3==1.1.0
 opencv_python==4.5.1.48
-gym==0.18.0
-numpy==1.19.5
+matplotlib==3.4.2
 dataclasses==0.8
-Pillow==8.3.2
-pre-commit==2.15.0
+Pillow==8.4.0
